Some Spring Ball tweaks
- Added cooldown reset for Spring Ball on morph/unmorph :

Fixes a bug where the cooldown would not reset because of unmorphing when the cooldown is still active.

- Nerfed Life Grove Tunnel climbing with Spring Ball :

It was completely unintended that the player could climb Life Grove Tunnel without bombs. As such a check was added so when the player tries to jump while rolling in a half pipe it decreases the vertical speed by 1.5x.

- Added fdivs command in ppcasm module :

It was required to divide the vertical speed which is used for nerfing Life Grove Tunnel Spring Ball climb.