Fixes seed hash not being shown on PAL main menu
Seed hash is not shown on PAL main menu when starting the game for the first time in a non english language.
This will remain the thing until the save file is deleted on memory card (not in main menu).
This commit fixes the above.

Link to bug report: [#bug-reports -> Seed doesn't display on main menu on PAL](https://discord.com/channels/914291389293027329/1175534362914193419)

Any other changes are just trimming whitespaces. It's done automatically via Notepad++ with Alt+Shift+S to save the file.